Add abstract base class Renderer
