+++ lib/contourpy/util/mpl_renderer.py
@@ -7,9 +7,10 @@ import numpy as np
 from contourpy import FillType, LineType
 
 from .mpl_util import filled_to_mpl_paths, lines_to_mpl_paths, mpl_codes_to_offsets
+from .renderer import Renderer
 
 
-class MplRenderer:
+class MplRenderer(Renderer):
     """Utility renderer using Matplotlib to render a grid of plots over the same (x, y) range.
 
     Args:
@@ -62,13 +63,6 @@ class MplRenderer:
         if isinstance(ax, int):
             ax = self._axes[ax]
         return ax
-
-    def _grid_as_2d(self, x, y):
-        x = np.asarray(x)
-        y = np.asarray(y)
-        if x.ndim == 1:
-            x, y = np.meshgrid(x, y)
-        return x, y
 
     def filled(self, filled, fill_type, ax=0, color="C0", alpha=0.7):
         """Plot filled contours on a single Axes.
